DoctorsController now uses database; improved doctors page

+++ src/main/java/ru/bert7438/stomatology/repositories/DoctorRepository.java
@@ -4,10 +4,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.bert7438.stomatology.entities.Doctor;
 
+import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface DoctorRepository extends JpaRepository<Doctor, Long> {
+public interface DoctorRepository extends JpaRepository<Doctor, Integer> {
 
     Optional<Doctor> findById(int id);
 }
